Moved linters of built-in subs
