+++ src/built_ins/input.rs
@@ -35,11 +35,52 @@ pub mod parser {
 }
 
 pub mod linter {
-    use crate::common::{QError, QErrorNode, ToErrorEnvelopeNoPos};
+    use crate::common::*;
     use crate::linter::arg_validation::ArgValidation;
-    use crate::parser::ExpressionNode;
+    use crate::parser::{Expression, ExpressionNode};
 
     pub fn lint(args: &Vec<ExpressionNode>) -> Result<(), QErrorNode> {
+        // the first one or two arguments stand for the file number
+        // if the first argument is 0, no file handle
+        // if the first argument is 1, the second is the file handle
+
+        if args.len() <= 1 {
+            return Err(QError::ArgumentCountMismatch).with_err_no_pos();
+        }
+        let mut has_file_number: bool = false;
+        if let Locatable {
+            element: Expression::IntegerLiteral(0),
+            ..
+        } = args[0]
+        {
+            // does not have a file number
+        } else if let Locatable {
+            element: Expression::IntegerLiteral(1),
+            ..
+        } = args[0]
+        {
+            // must have a file number
+            args.require_integer_argument(1)?;
+            has_file_number = true;
+        } else {
+            panic!("parser sent unexpected arguments");
+        }
+
+        let starting_index = if has_file_number { 2 } else { 1 };
+        if args.len() <= starting_index {
+            return Err(QError::ArgumentCountMismatch).with_err_no_pos();
+        }
+
+        for i in starting_index..args.len() {
+            let Locatable { element, .. } = &args[i];
+            match element {
+                Expression::Variable(_, _) | Expression::Property(_, _, _) => {}
+                _ => {
+                    return Err(QError::VariableRequired).with_err_at(&args[i]);
+                }
+            }
+        }
+
         Ok(())
     }
 }
@@ -56,6 +97,7 @@ pub mod interpreter {
 #[cfg(test)]
 mod tests {
     use crate::assert_built_in_sub_call;
+    use crate::assert_linter_err;
     use crate::assert_parser_err;
     use crate::built_ins::BuiltInSub;
     use crate::common::*;
@@ -137,5 +179,26 @@ mod tests {
             Expression::IntegerLiteral(3), // file number
             Expression::var_unresolved("A")
         );
+    }
+
+    #[test]
+    fn test_parenthesis_variable_required() {
+        let input = "INPUT (A$)";
+        assert_linter_err!(input, QError::VariableRequired);
+    }
+
+    #[test]
+    fn test_binary_expression_variable_required() {
+        let input = "INPUT A$ + B$";
+        assert_linter_err!(input, QError::VariableRequired);
+    }
+
+    #[test]
+    fn test_const() {
+        let input = r#"
+                CONST A$ = "hello"
+                INPUT A$
+                "#;
+        assert_linter_err!(input, QError::VariableRequired);
     }
 }
