+++ src/built_ins/open.rs
@@ -152,6 +152,19 @@ pub mod linter {
     use crate::parser::ExpressionNode;
 
     pub fn lint(args: &Vec<ExpressionNode>) -> Result<(), QErrorNode> {
+        // must have 5 arguments:
+        // filename
+        // file mode
+        // file access
+        // file number
+        // rec len
+        if args.len() != 5 {
+            return Err(QError::ArgumentCountMismatch).with_err_no_pos();
+        }
+        args.require_string_argument(0)?;
+        for i in 1..args.len() {
+            args.require_integer_argument(i)?;
+        }
         Ok(())
     }
 }
@@ -167,6 +180,7 @@ pub mod interpreter {
 
 #[cfg(test)]
 mod tests {
+    use crate::assert_linter_err;
     use crate::assert_parser_err;
     use crate::built_ins::BuiltInSub;
     use crate::common::*;
@@ -334,5 +348,17 @@ mod tests {
                 ]
             )
         );
+    }
+
+    #[test]
+    fn test_open_filename_must_be_string() {
+        let program = "OPEN 42 AS #1";
+        assert_linter_err!(program, QError::ArgumentTypeMismatch, 1, 6);
+    }
+
+    #[test]
+    fn test_rec_len_must_be_numeric() {
+        let program = r#"OPEN "a.txt" AS #1 LEN = "hi""#;
+        assert_linter_err!(program, QError::ArgumentTypeMismatch, 1, 26);
     }
 }
