+++ src/linter/post_linter/built_in_linter.rs
@@ -14,20 +14,7 @@ impl PostConversionLinter for BuiltInLinter {
         args: &Vec<ExpressionNode>,
     ) -> Result<(), QErrorNode> {
         self.visit_expressions(args)?;
-        match built_in_sub {
-            BuiltInSub::Close => close::lint(args),
-            BuiltInSub::Environ => environ_sub::lint(args),
-            BuiltInSub::Field => field::lint(args),
-            BuiltInSub::Get => get::lint(args),
-            BuiltInSub::Input => input::lint(args),
-            BuiltInSub::Kill => kill::lint(args),
-            BuiltInSub::LineInput => line_input::lint(args),
-            BuiltInSub::LSet => lset::lint(args),
-            BuiltInSub::Name => name::lint(args),
-            BuiltInSub::Open => open::lint(args),
-            BuiltInSub::Put => put::lint(args),
-            BuiltInSub::ViewPrint => crate::built_ins::view_print::linter::lint(args),
-        }
+        crate::built_ins::linter::lint_sub_call(built_in_sub, args)
     }
 
     fn visit_expression(&mut self, expr_node: &ExpressionNode) -> Result<(), QErrorNode> {
@@ -61,160 +48,6 @@ fn lint(built_in: &BuiltInFunction, args: &Vec<ExpressionNode>) -> Result<(), QE
         BuiltInFunction::Mid => mid::lint(args),
         BuiltInFunction::Str => str_fn::lint(args),
         BuiltInFunction::Val => val::lint(args),
-    }
-}
-
-mod close {
-    use super::*;
-    use crate::linter::arg_validation::ArgValidation;
-
-    pub fn lint(args: &Vec<ExpressionNode>) -> Result<(), QErrorNode> {
-        for i in 0..args.len() {
-            args.require_integer_argument(i)?;
-        }
-        Ok(())
-    }
-}
-
-mod environ_sub {
-    use super::*;
-
-    pub fn lint(args: &Vec<ExpressionNode>) -> Result<(), QErrorNode> {
-        if args.len() != 1 {
-            Err(QError::ArgumentCountMismatch).with_err_no_pos()
-        } else if !args[0].can_cast_to(TypeQualifier::DollarString) {
-            Err(QError::ArgumentTypeMismatch).with_err_at(&args[0])
-        } else {
-            Ok(())
-        }
-    }
-}
-
-mod field {
-    use super::*;
-    use crate::linter::arg_validation::ArgValidation;
-
-    pub fn lint(args: &Vec<ExpressionNode>) -> Result<(), QErrorNode> {
-        // needs to be 1 + N*3 args, N >= 1
-        // first is the file number
-        // then the fields: width, variable name, variable
-        if args.len() < 4 {
-            return Err(QError::ArgumentCountMismatch).with_err_no_pos();
-        }
-        if (args.len() - 1) % 3 != 0 {
-            return Err(QError::ArgumentCountMismatch).with_err_no_pos();
-        }
-        args.require_integer_argument(0)?;
-        let mut i: usize = 1;
-        while i < args.len() {
-            args.require_integer_argument(i)?;
-            args.require_string_argument(i + 1)?;
-            args.require_string_variable(i + 2)?;
-            i += 3;
-        }
-        Ok(())
-    }
-}
-
-mod get {
-    use super::*;
-    use crate::linter::arg_validation::ArgValidation;
-
-    pub fn lint(args: &Vec<ExpressionNode>) -> Result<(), QErrorNode> {
-        if args.len() != 2 {
-            return Err(QError::ArgumentCountMismatch).with_err_no_pos();
-        }
-        args.require_integer_argument(0)?;
-        args.require_long_argument(1)
-    }
-}
-
-mod input {
-    use crate::common::ToErrorEnvelopeNoPos;
-
-    use super::*;
-    use crate::linter::arg_validation::ArgValidation;
-
-    pub fn lint(args: &Vec<ExpressionNode>) -> Result<(), QErrorNode> {
-        // the first one or two arguments stand for the file number
-        // if the first argument is 0, no file handle
-        // if the first argument is 1, the second is the file handle
-
-        if args.len() <= 1 {
-            return Err(QError::ArgumentCountMismatch).with_err_no_pos();
-        }
-        let mut has_file_number: bool = false;
-        if let Locatable {
-            element: Expression::IntegerLiteral(0),
-            ..
-        } = args[0]
-        {
-            // does not have a file number
-        } else if let Locatable {
-            element: Expression::IntegerLiteral(1),
-            ..
-        } = args[0]
-        {
-            // must have a file number
-            args.require_integer_argument(1)?;
-            has_file_number = true;
-        } else {
-            panic!("parser sent unexpected arguments");
-        }
-
-        let starting_index = if has_file_number { 2 } else { 1 };
-        if args.len() <= starting_index {
-            return Err(QError::ArgumentCountMismatch).with_err_no_pos();
-        }
-
-        for i in starting_index..args.len() {
-            let Locatable { element, .. } = &args[i];
-            match element {
-                Expression::Variable(_, _) | Expression::Property(_, _, _) => {}
-                _ => {
-                    return Err(QError::VariableRequired).with_err_at(&args[i]);
-                }
-            }
-        }
-
-        Ok(())
-    }
-
-    #[cfg(test)]
-    mod tests {
-        use crate::assert_linter_err;
-
-        use super::*;
-
-        #[test]
-        fn test_parenthesis_variable_required() {
-            let input = "INPUT (A$)";
-            assert_linter_err!(input, QError::VariableRequired);
-        }
-
-        #[test]
-        fn test_binary_expression_variable_required() {
-            let input = "INPUT A$ + B$";
-            assert_linter_err!(input, QError::VariableRequired);
-        }
-
-        #[test]
-        fn test_const() {
-            let input = r#"
-                CONST A$ = "hello"
-                INPUT A$
-                "#;
-            assert_linter_err!(input, QError::VariableRequired);
-        }
-    }
-}
-
-mod kill {
-    use super::*;
-    use crate::linter::arg_validation::ArgValidation;
-
-    pub fn lint(args: &Vec<ExpressionNode>) -> Result<(), QErrorNode> {
-        args.require_one_string_argument()
     }
 }
 
@@ -252,155 +85,6 @@ mod lbound {
         } else {
             Err(QError::ArgumentTypeMismatch).with_err_at(first_pos)
         }
-    }
-}
-
-mod line_input {
-    use crate::common::ToErrorEnvelopeNoPos;
-
-    use super::*;
-
-    pub fn lint(args: &Vec<ExpressionNode>) -> Result<(), QErrorNode> {
-        // the first one or two arguments stand for the file number
-        // if the first argument is 0, no file handle
-        // if the first argument is 1, the second is the file handle
-
-        if args.len() <= 1 {
-            return Err(QError::ArgumentCountMismatch).with_err_no_pos();
-        }
-        let mut has_file_number: bool = false;
-        if let Locatable {
-            element: Expression::IntegerLiteral(0),
-            ..
-        } = args[0]
-        {
-            // does not have a file number
-        } else if let Locatable {
-            element: Expression::IntegerLiteral(1),
-            ..
-        } = args[0]
-        {
-            // must have a file number
-            if let Locatable {
-                element: Expression::IntegerLiteral(_),
-                ..
-            } = args[1]
-            {
-                has_file_number = true;
-            } else {
-                panic!("parser sent unexpected arguments");
-            }
-        } else {
-            panic!("parser sent unexpected arguments");
-        }
-
-        let starting_index = if has_file_number { 2 } else { 1 };
-        if args.len() != starting_index + 1 {
-            return Err(QError::ArgumentCountMismatch).with_err_no_pos();
-        }
-
-        let Locatable {
-            element: var_arg,
-            pos,
-        } = &args[starting_index];
-        match var_arg {
-            Expression::Variable(
-                _,
-                VariableInfo {
-                    expression_type, ..
-                },
-            ) => match expression_type {
-                ExpressionType::BuiltIn(TypeQualifier::DollarString)
-                | ExpressionType::FixedLengthString(_) => {}
-                _ => return Err(QError::TypeMismatch).with_err_at(*pos),
-            },
-            _ => return Err(QError::TypeMismatch).with_err_at(*pos),
-        }
-
-        Ok(())
-    }
-}
-
-mod lset {
-    use super::*;
-    use crate::linter::arg_validation::ArgValidation;
-
-    pub fn lint(args: &Vec<ExpressionNode>) -> Result<(), QErrorNode> {
-        // the parser should produce 3 arguments:
-        // the variable name, as a string literal
-        // the variable itself, a ByRef string variable
-        // a string expression to assign to
-        if args.len() != 3 {
-            return Err(QError::ArgumentCountMismatch).with_err_no_pos();
-        }
-        args.require_string_argument(0)?;
-        // TODO ensure LSET is operating on variables previously used by FIELD in this scope
-        args.require_string_variable(1)?;
-        args.require_string_argument(2)
-    }
-}
-
-mod name {
-    use super::*;
-
-    pub fn lint(args: &Vec<ExpressionNode>) -> Result<(), QErrorNode> {
-        if args.len() != 2 {
-            Err(QError::ArgumentCountMismatch).with_err_no_pos()
-        } else if !args[0].can_cast_to(TypeQualifier::DollarString) {
-            Err(QError::ArgumentTypeMismatch).with_err_at(&args[0])
-        } else if !args[1].can_cast_to(TypeQualifier::DollarString) {
-            Err(QError::ArgumentTypeMismatch).with_err_at(&args[1])
-        } else {
-            Ok(())
-        }
-    }
-}
-
-mod open {
-    use super::*;
-    use crate::linter::arg_validation::ArgValidation;
-
-    pub fn lint(args: &Vec<ExpressionNode>) -> Result<(), QErrorNode> {
-        // must have 5 arguments:
-        // filename
-        // file mode
-        // file access
-        // file number
-        // rec len
-        if args.len() != 5 {
-            return Err(QError::ArgumentCountMismatch).with_err_no_pos();
-        }
-        args.require_string_argument(0)?;
-        for i in 1..args.len() {
-            args.require_integer_argument(i)?;
-        }
-        Ok(())
-    }
-
-    #[cfg(test)]
-    mod tests {
-        use super::*;
-        use crate::assert_linter_err;
-
-        #[test]
-        fn test_open_filename_must_be_string() {
-            let program = "OPEN 42 AS #1";
-            assert_linter_err!(program, QError::ArgumentTypeMismatch, 1, 6);
-        }
-
-        #[test]
-        fn test_rec_len_must_be_numeric() {
-            let program = r#"OPEN "a.txt" AS #1 LEN = "hi""#;
-            assert_linter_err!(program, QError::ArgumentTypeMismatch, 1, 26);
-        }
-    }
-}
-
-mod put {
-    use super::*;
-
-    pub fn lint(args: &Vec<ExpressionNode>) -> Result<(), QErrorNode> {
-        super::get::lint(args)
     }
 }
